Handle account alias properly in accounts command.
This also comes with refactoring splitting
book keep details (add_transaction) and process overall,
so that we can reuse it from accounts.
